Added current files
rhox1d, fluxes1D, and evolve1D are helper functions for test1D, which is just to compute plots for various examples. rhox2d and rhoy2d, both support test2D to generate the plots in Section 4.3. Minmod helps all of the above. Test1Dconvergence and test1Dentropy are self contained tests of the convergence to a known steady state in 1D as well as a test of the convergence of entropy to a minimum for a compactly supported kernel in 1D.